Extract go to definition
