+++ __init__.py
@@ -4,4 +4,5 @@ from .features.completion import *
 from .features.diagnostics import *
 from .features.configuration import *
 from .features.formatting import *
+from .features.definition import *
 from .features.hover import *
