+++ core/main.py
@@ -158,42 +158,6 @@ def preserve_whitespace(contents: str) -> str:
     return contents
 
 
-class LspSymbolDefinitionCommand(sublime_plugin.TextCommand):
-    def is_enabled(self, event=None):
-        # TODO: check what kind of scope we're in.
-        if is_supported_view(self.view):
-            client = client_for_view(self.view)
-            if client and client.has_capability('definitionProvider'):
-                return is_at_word(self.view, event)
-        return False
-
-    def run(self, edit, event=None):
-        client = client_for_view(self.view)
-        if client:
-            pos = get_position(self.view, event)
-            document_position = get_document_position(self.view, pos)
-            if document_position:
-                request = Request.definition(document_position)
-                client.send_request(
-                    request, lambda response: self.handle_response(response, pos))
-
-    def handle_response(self, response, position):
-        window = sublime.active_window()
-        if response:
-            location = response if isinstance(response, dict) else response[0]
-            file_path = uri_to_filename(location.get("uri"))
-            start = Point.from_lsp(location['range']['start'])
-            file_location = "{}:{}:{}".format(file_path, start.row + 1, start.col + 1)
-            debug("opening location", location)
-            window.open_file(file_location, sublime.ENCODED_POSITION)
-            # TODO: can add region here.
-        else:
-            window.run_command("goto_definition")
-
-    def want_event(self):
-        return True
-
-
 def format_symbol_kind(kind):
     return symbol_kind_names.get(kind, str(kind))
